PR comments

+++ src/pheromones/hivemind-host/include/hivemind-host/FunctionCallArgumentDTO.h
@@ -13,7 +13,7 @@ class FunctionCallArgumentDTO {
   public:
     FunctionCallArgumentDTO(const FunctionArgument& argument);
 
-    FunctionCallArgumentDTO(int32_t argument);
+    FunctionCallArgumentDTO(int64_t argument);
 
     /**
      *@brief default constructor, the argument will be initialized as a monostate (ie, no arguments)
@@ -24,14 +24,14 @@ class FunctionCallArgumentDTO {
      *@brief get the current stored argument
      *
      *@return the argument*/
-    const std::variant<std::monostate, int32_t>& getArgument() const;
+    const std::variant<std::monostate, int64_t>& getArgument() const;
 
     /**
      *@brief set the stored argument
      *
      *@param [in] arg the value to set
      */
-    void setArgument(const std::variant<std::monostate, int32_t>& arg);
+    void setArgument(const std::variant<std::monostate, int64_t>& arg);
 
     /**
      *@brief serialize a FunctionArgument for nanopb, sets the fields properly before using
@@ -44,7 +44,7 @@ class FunctionCallArgumentDTO {
     bool serialize(FunctionArgument& argument) const;
 
   private:
-    std::variant<std::monostate, int32_t> m_argument;
+    std::variant<std::monostate, int64_t> m_argument;
 };
 
 #endif // __FUNCTIONCALLARGUMENTDTO_H_
