+++ src/pheromones/hivemind-host/src/FunctionCallArgumentDTO.cpp
@@ -2,7 +2,7 @@
 
 FunctionCallArgumentDTO::FunctionCallArgumentDTO() { m_argument = std::monostate(); }
 
-FunctionCallArgumentDTO::FunctionCallArgumentDTO(int32_t argument) { m_argument = argument; }
+FunctionCallArgumentDTO::FunctionCallArgumentDTO(int64_t argument) { m_argument = argument; }
 
 FunctionCallArgumentDTO::FunctionCallArgumentDTO(const FunctionArgument& argument) {
     switch (argument.which_argument) {
@@ -15,16 +15,16 @@ FunctionCallArgumentDTO::FunctionCallArgumentDTO(const FunctionArgument& argumen
     }
 }
 
-const std::variant<std::monostate, int32_t>& FunctionCallArgumentDTO::getArgument() const {
+const std::variant<std::monostate, int64_t>& FunctionCallArgumentDTO::getArgument() const {
     return m_argument;
 }
 
-void FunctionCallArgumentDTO::setArgument(const std::variant<std::monostate, int32_t>& arg) {
+void FunctionCallArgumentDTO::setArgument(const std::variant<std::monostate, int64_t>& arg) {
     m_argument = arg;
 }
 
 bool FunctionCallArgumentDTO::serialize(FunctionArgument& argument) const {
-    if (const int32_t* intArg = std::get_if<int32_t>(&m_argument)) {
+    if (const int64_t* intArg = std::get_if<int64_t>(&m_argument)) {
         argument.which_argument = FunctionArgument_int_arg_tag;
         argument.argument.int_arg = *intArg;
         return true;
